+++ setup.py
@@ -10,9 +10,18 @@ setup(name='shakemap',
       author='Bruce Worden, Mike Hearne, Eric Thompson',
       author_email='cbworden@usgs.gov,mhearne@usgs.gov,emthompson@usgs.gov',
       url='http://github.com/usgs/shakemap',
-      packages=['shakemap', 'shakemap.gmpe', 'shakemap.grind', 'shakemap.genex',
-                'shakemap.mapping', 'shakemap.utils', 'shakemap.correlation',
-                'shakemap.gmice', 'shakemap.directivity', 'shakemap.transfer',
+      packages=['shakemap', 
+                'shakemap.grind', 
+                'shakemap.grind.conversions', 
+                'shakemap.grind.conversions.imc', 
+                'shakemap.grind.conversions.imt', 
+                'shakemap.grind.correlation', 
+                'shakemap.grind.directivity', 
+                'shakemap.genex',
+                'shakemap.mapping', 
+                'shakemap.utils', 
+                'shakemap.gmice', 
+                'shakemap.transfer',
                 'shakemap.extern.scp'],
       package_data={'shakemap': [os.path.join('utils', 'configspec.ini'),
                                  os.path.join('grind', 'data', 'ps2ff', '*.csv')]},
