+++ shakemap/grind/correlation/goda_atkinson_2010.py
