+++ shakemap/grind/directivity/bayless2013.py
@@ -5,10 +5,10 @@ import openquake.hazardlib.geo as geo
 import copy
 
 import shakemap.grind.fault as fault
-import shakemap.grind.ecef as ecef
 from shakemap.grind.vector import Vector
 from shakemap.grind.distance import get_distance
 from shakemap.grind.distance import distance_sq_to_segment
+import shakemap.utils.ecef as ecef
 
 
 """
