+++ shakemap/grind/gmpe2shakemap.py
@@ -4,7 +4,7 @@ GMPE amps are in ln(g), in an IMC of their choosing
 ShakeMap amps are %g and max horizontal
 """
 
-from . import BeyerBommer2006 as BB
+from shakemap.grind.conversions.imc.beyer_bommer_2006 import BeyerBommer2006 as bb06
 
 from openquake.hazardlib.imt import PGA, PGV, SA
 from openquake.hazardlib import const
@@ -17,7 +17,7 @@ def ampShakeMapToGMPE(sm_amps, gmpe, imt):
     else:
         gmpe_amps = sm_amps.copy()
 
-    gmpe_amps = BB.ampIMCtoIMC(gmpe_amps, const.IMC.GREATER_OF_TWO_HORIZONTAL,
+    gmpe_amps = bb06.ampIMCtoIMC(gmpe_amps, const.IMC.GREATER_OF_TWO_HORIZONTAL,
                                gmpe.DEFINED_FOR_INTENSITY_MEASURE_COMPONENT, imt)
     gmpe_amps = gmpe.to_distribution_values(gmpe_amps)
 
@@ -27,7 +27,7 @@ def ampShakeMapToGMPE(sm_amps, gmpe, imt):
 def ampGmpeToShakeMap(gmpe_amps, gmpe, imt):
 
     sm_amps = gmpe.to_imt_unit_values(gmpe_amps)
-    sm_amps = BB.ampIMCtoIMC(sm_amps, gmpe.DEFINED_FOR_INTENSITY_MEASURE_COMPONENT,
+    sm_amps = bb06.ampIMCtoIMC(sm_amps, gmpe.DEFINED_FOR_INTENSITY_MEASURE_COMPONENT,
                              const.IMC.GREATER_OF_TWO_HORIZONTAL, imt)
 
     if not isinstance(imt, PGV):
@@ -39,7 +39,7 @@ def ampGmpeToShakeMap(gmpe_amps, gmpe, imt):
 def sigmaShakeMapToGMPE(sm_sigmas, gmpe, imt):
 
     gmpe_sigmas = gmpe.to_distribution_values(sm_sigmas)
-    gmpe_sigmas = BB.sigmaIMCtoIMC(gmpe_sigmas, const.IMC.GREATER_OF_TWO_HORIZONTAL,
+    gmpe_sigmas = bb06.sigmaIMCtoIMC(gmpe_sigmas, const.IMC.GREATER_OF_TWO_HORIZONTAL,
                                    gmpe.DEFINED_FOR_INTENSITY_MEASURE_COMPONENT, imt)
 
     return gmpe_sigmas
@@ -47,7 +47,7 @@ def sigmaShakeMapToGMPE(sm_sigmas, gmpe, imt):
 
 def sigmaGmpeToShakeMap(gmpe_sigmas, gmpe, imt):
 
-    sm_sigmas = BB.sigmaIMCtoIMC(gmpe_sigmas, gmpe.DEFINED_FOR_INTENSITY_MEASURE_COMPONENT,
+    sm_sigmas = bb06.sigmaIMCtoIMC(gmpe_sigmas, gmpe.DEFINED_FOR_INTENSITY_MEASURE_COMPONENT,
                                  const.IMC.GREATER_OF_TWO_HORIZONTAL, imt)
     sm_sigmas = gmpe.to_imt_unit_values(sm_sigmas)
