+++ shakemap/grind/vector.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 from openquake.hazardlib.geo import point
-from .ecef import latlon2ecef, ecef2latlon
+from ..utils.ecef import latlon2ecef, ecef2latlon
 
 
 class Vector(object):
